requirements.txt
hydra-core==1.3.2
lightning==2.4.0
lightning_utilities==0.11.8
matplotlib==3.10.0
numpy==2.2.3
omegaconf==2.3.0
pandas==2.2.3
ptflops==0.7.4
PyYAML==6.0.2
PyYAML==6.0.2
scikit_learn==1.6.1
thop==0.1.1.post2209072238
torch==2.5.1
